Move polyfills into utils/
